requirements.txt


fastf1
psycopg2-binary
pandas
prefect[dotenv]
plotly
dash
python-dotenv
sqlalchemy